% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vegas.R
\name{get_matchups}
\alias{get_matchups}
\title{Scrape a week of NFL matchups for a given season from CBS.com}
\usage{
get_matchups(year = NULL, week = NULL, verbose = TRUE)
}
\arguments{
\item{year}{numeric for season of interest}

\item{week}{numeric for the week that we want to get the team matchups for}

\item{verbose}{logical, whether to print messages as function runs. Default is to print messages (TRUE)}
}
\value{
dataframe with the home and away teams for the desired week
}
\description{
Finds the team matchups for a specific week of an NFL season
}
\examples{
# Get NFL matchups for a given season and week
matchups <- get_matchups(
                        year = 2022,
                        week = 1
                       )
matchups
}
