% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nfl_elo}
\alias{get_nfl_elo}
\title{Calculate NFL Elo Ratings}
\usage{
get_nfl_elo(nfl_season)
}
\arguments{
\item{nfl_season}{dataframe with home/away teams, week, home/away team scores, and game outcomes}
}
\value{
dataframe
}
\description{
Calculates Elo ratings for an NFL season
}
