% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_teams.R
\name{get_nfl_teams}
\alias{get_nfl_teams}
\title{Get NFL teams and team abbreviations}
\usage{
get_nfl_teams()
}
\value{
tibble of NFL teams and team abbreviations
}
\description{
Returns a tibble with all 32 NFL teams and team abbreviations. Includes more than 32 team names due to teams changing names and/or cities.
}
