% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_plot.R
\name{get_prob_plot}
\alias{get_prob_plot}
\title{Plot Prediction Probability}
\usage{
get_prob_plot(predictions = NULL, pct_over_coinflip = FALSE)
}
\arguments{
\item{predictions}{dataframe with .pred_class, .pred_1, .pred_0 and home_team columns. Output from predict_games() function}

\item{pct_over_coinflip}{logical. If TRUE, probabilities are returned to represent how far over a 50-50 guess the prediction is. For example, a value of 0 on the plot x axis would mean that the home and away team have a 50\% chance of winning the game.}
}
\value{
ggplot object with home teams on the y axis and probabilities of the correct prediction on the x axis
}
\description{
Takes the output predictions from predict_games() and plots the predicted outcomes for the home teams
}
