% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vegas.R
\name{get_vegas}
\alias{get_vegas}
\title{Scrape internet for Vegas Odds for current NFL season}
\usage{
get_vegas(current_week = TRUE)
}
\arguments{
\item{current_week}{logical, whether to return all available odds or just the current weeks odds. Default is TRUE, returns just the current week odds}
}
\value{
tibble with a row for each game with the odds columns from the perspective of the home team (spread, win, over under total)
}
\description{
Returns a dataframe with NFL odds for either the current week of the NFL season, or all avaliable weeks.
}
