% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_week.R
\name{get_week}
\alias{get_week}
\title{Find the week of the NFL season according to a date}
\usage{
get_week(date = NULL)
}
\arguments{
\item{date}{character date, YYYY-MM-DD. Default is NULL and date will set to the date the function was run}
}
\value{
character string of the current week of the NFL season
}
\description{
returns week of the NFL season by a date, if no date is entered, default is to return the current week of the season.
}
