% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_week_dates.R
\name{get_week_dates}
\alias{get_week_dates}
\title{Find dates coorsponding to weeks of a given NFL season}
\usage{
get_week_dates(date = NULL)
}
\arguments{
\item{date}{character date, YYYY-MM-DD. Default is NULL and date will set to the date the function was run (i.e. the current season)}
}
\value{
tibble showing the weeks of the desired NFL season with start_week and end_week columns indicating the dates that a week starts and ends, respectively.
}
\description{
Returns a tibble with the starting and ending dates of each week in a given NFL season.
}
