% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_win_pct}
\alias{get_win_pct}
\title{Calculate win percentages}
\usage{
get_win_pct(df, verbose = TRUE)
}
\arguments{
\item{df}{dataframe with home/away teams, week, and game outcome}

\item{verbose}{logical, if TRUE, prints log message. Default is TRUE}
}
\value{
dataframe
}
\description{
Calculates overall, home, and away win percentages
}
