% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R
\name{get_year}
\alias{get_year}
\title{Find the year of the NFL season for a given date}
\usage{
get_year(date = NULL)
}
\arguments{
\item{date}{character string date, default is NULL and date will set to the date the function was run}
}
\value{
numeric indicating the season year
}
\description{
Returns numeric year of the NFL season for a given date
}
