% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prob.R
\name{plot_prob}
\alias{plot_prob}
\title{Plot Prediction Probability}
\usage{
plot_prob(
  predictions = NULL,
  bar_plot = FALSE,
  pct_over_coinflip = FALSE,
  prob_alpha = TRUE
)
}
\arguments{
\item{predictions}{dataframe with .pred_class, .pred_1, .pred_0 and home_team columns. Output from predict_games() function}

\item{bar_plot}{logical. If TRUE, a bar plot is generated. Default is FALSE and will return a lollipop plot}

\item{pct_over_coinflip}{logical. If TRUE, probabilities are returned to represent how far over a 50-50 guess the prediction is. For example, a value of 0 on the plot x axis would mean that the home and away team have a 50\% chance of winning the game.}

\item{prob_alpha}{Whether the lollipop or bars on plot should be more/less transparent based on the model class probability. Default is TRUE, alpha is used on plot elements}
}
\value{
ggplot object with team matchups on the y axis and probabilities of the correct prediction on the x axis
}
\description{
Takes the output predictions from predict_games() and plots the predicted outcomes for the home teams
}
