% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tile.R
\name{plot_tile}
\alias{plot_tile}
\title{Plot home and away win probabilities on a grid for a given week}
\usage{
plot_tile(predictions = NULL, favored_only = FALSE)
}
\arguments{
\item{predictions}{dataframe with win, home_win_prob, away_win_prob and home_team columns. Output from predict_games() function}

\item{favored_only}{logical. whether to only plot the favored teams win probabilities, or both team outcomes. Default is FALSE, to return win probability for both teams.}
}
\value{
ggplot object with matchups on the Y axis and the X axis split by home and away teams. Darker red colors indicate a lower win probability and dark green colors indicate a higher win probability.
}
\description{
Takes the output predictions from predict_games() and plots the prediction probabilities for both teams for a given week.
}
