% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_games.R
\name{predict_games}
\alias{predict_games}
\title{Generate NFL win predictions for past and upcoming NFL games}
\usage{
predict_games(year = NULL, week = NULL)
}
\arguments{
\item{year}{character or numeric year of season to get predictions for}

\item{week}{numeric value for the week of the season to predict}
}
\value{
tibble with a row for each matchup between teams for the desired week. There is basic identifying information on each game and three columns representing the model predicted outcomes: .pred_class, .pred_1, .pred_0. .pred_class gives a 1 if the home team is predicted to win, and a 0 if the home team is predicted to lose. The .pred_1 and .pred_0 columns indicate the probability of either class occurring.
}
\description{
Retrieve NFL game predictions for a given year and week
}
\examples{
# Predict week 8 of the 2021 NFL season
predict_games(year = 2021, week = 8)
}
