% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_page}
\alias{process_page}
\title{Process HTML table nodes from profootballreference.com}
\usage{
process_page(page)
}
\arguments{
\item{page}{xml_document created by using rvest::read_html on desired page}
}
\value{
data.frame
}
\description{
Cleans pages with game details from profootballreference.com
}
