% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_games}
\alias{scrape_games}
\title{Scrape data needed to make a prediction for a given week}
\usage{
scrape_games(year = NULL, pred_week = NULL)
}
\arguments{
\item{year}{numeric for season of interest}

\item{pred_week}{numeric for the week that we want to get the team matchups for}
}
\value{
dataframe with necessary model inputs
}
\description{
Retrieves the necessary model input data from profootballreference.com and returns a dataframe that can be inputted into models
}
