% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_model_data.R
\docType{data}
\name{win_model}
\alias{win_model}
\title{Logistic Regression Workflow object}
\format{
An object of class \code{workflow} of length 4.
}
\usage{
win_model
}
\description{
Logistic Regression model used to predict NFL wins.
}
\examples{
\dontrun{
 nflpredictr::win_model
}

}
\keyword{datasets}
