% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_games.R
\name{predict_games}
\alias{predict_games}
\title{Request predictions from NFL Win Predictor API}
\usage{
predict_games(
  year = NULL,
  week = NULL,
  base_url = "http://68.183.25.9:8000/predict-new-data?"
)
}
\arguments{
\item{year}{character or numeric year of season to get predictions for}

\item{week}{numeric value for the week of the season to predict}

\item{base_url}{character string defining the base API endpoint URL. Default is predict-new-data endpoint}
}
\value{
numeric indicating the season year
}
\description{
Returns numeric year of the NFL season for a given date
}
